+++ core/src/main/java/cn/edu/tsinghua/iotdb/benchmark/entity/Batch.java
@@ -17,9 +17,9 @@
  * under the License.
  */
 
-package cn.edu.tsinghua.iotdb.benchmark.workload.ingestion;
+package cn.edu.tsinghua.iotdb.benchmark.entity;
 
-import cn.edu.tsinghua.iotdb.benchmark.schema.enums.SensorType;
+import cn.edu.tsinghua.iotdb.benchmark.entity.enums.SensorType;
 import cn.edu.tsinghua.iotdb.benchmark.schema.schemaImpl.DeviceSchema;
 import cn.edu.tsinghua.iotdb.benchmark.utils.ReadWriteIOUtils;
 import org.apache.commons.lang3.builder.EqualsBuilder;
