+++ core/src/main/java/cn/edu/tsinghua/iotdb/benchmark/kafka/BatchConsumer.java
@@ -21,7 +21,7 @@ package cn.edu.tsinghua.iotdb.benchmark.kafka;
 
 import cn.edu.tsinghua.iotdb.benchmark.conf.Config;
 import cn.edu.tsinghua.iotdb.benchmark.conf.ConfigDescriptor;
-import cn.edu.tsinghua.iotdb.benchmark.workload.ingestion.Batch;
+import cn.edu.tsinghua.iotdb.benchmark.entity.Batch;
 import kafka.consumer.KafkaStream;
 import kafka.javaapi.consumer.ConsumerConnector;
 import kafka.serializer.StringDecoder;
