+++ core/src/main/java/cn/edu/tsinghua/iotdb/benchmark/kafka/BatchProducer.java
@@ -21,7 +21,7 @@ package cn.edu.tsinghua.iotdb.benchmark.kafka;
 
 import cn.edu.tsinghua.iotdb.benchmark.conf.Config;
 import cn.edu.tsinghua.iotdb.benchmark.conf.ConfigDescriptor;
-import cn.edu.tsinghua.iotdb.benchmark.workload.ingestion.Batch;
+import cn.edu.tsinghua.iotdb.benchmark.entity.Batch;
 import org.apache.kafka.clients.producer.KafkaProducer;
 import org.apache.kafka.clients.producer.Producer;
 import org.apache.kafka.clients.producer.ProducerRecord;
