+++ core/src/main/java/cn/edu/tsinghua/iotdb/benchmark/workload/interfaces/IGenerateDataWorkload.java
@@ -19,9 +19,9 @@
 
 package cn.edu.tsinghua.iotdb.benchmark.workload.interfaces;
 
+import cn.edu.tsinghua.iotdb.benchmark.entity.Batch;
+import cn.edu.tsinghua.iotdb.benchmark.exception.WorkloadException;
 import cn.edu.tsinghua.iotdb.benchmark.schema.schemaImpl.DeviceSchema;
-import cn.edu.tsinghua.iotdb.benchmark.workload.WorkloadException;
-import cn.edu.tsinghua.iotdb.benchmark.workload.ingestion.Batch;
 import cn.edu.tsinghua.iotdb.benchmark.workload.query.impl.*;
 
 /** IGenerateWorkload is a workload using for data which is generated by benchmark */
