+++ core/src/test/java/cn/edu/tsinghua/iotdb/benchmark/serialize/BatchSerializeTest.java
@@ -19,9 +19,9 @@
 
 package cn.edu.tsinghua.iotdb.benchmark.serialize;
 
+import cn.edu.tsinghua.iotdb.benchmark.entity.Batch;
+import cn.edu.tsinghua.iotdb.benchmark.entity.Record;
 import cn.edu.tsinghua.iotdb.benchmark.schema.schemaImpl.DeviceSchema;
-import cn.edu.tsinghua.iotdb.benchmark.workload.ingestion.Batch;
-import cn.edu.tsinghua.iotdb.benchmark.workload.ingestion.Record;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
